// src/generated/models/Afb_documentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_documentsafb_authentifie = {
  0: 'Oui',
  1: 'Non'
} as const;
export type Afb_documentsafb_authentifie = keyof typeof Afb_documentsafb_authentifie;
export const Afb_documentsafb_sourcedudepot = {
  747010000: 'D_pos_partiers',
  747010001: 'Ajout_parDCONF',
  747010002: 'G_n_r_parlesyst_me'
} as const;
export type Afb_documentsafb_sourcedudepot = keyof typeof Afb_documentsafb_sourcedudepot;
export const Afb_documentsafb_statutdevalidite = {
  0: 'Valide',
  1: 'Enattentedevalidation',
  747010001: 'Expir_',
  747010002: 'Rejet_'
} as const;
export type Afb_documentsafb_statutdevalidite = keyof typeof Afb_documentsafb_statutdevalidite;
export const Afb_documentsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_documentsstatecode = keyof typeof Afb_documentsstatecode;
export const Afb_documentsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_documentsstatuscode = keyof typeof Afb_documentsstatuscode;

export interface Afb_documentsBase {
  afb_anneededepot: number;
  afb_authentificateur?: string;
  afb_authentifie: Afb_documentsafb_authentifie;
  "afb_categorie@odata.bind": string;
  afb_datedemission?: string;
  afb_datedeteleversement: string;
  afb_datedexpiration?: string;
  afb_documentid: string;
  afb_motifderejet?: string;
  afb_nomdufichier: string;
  "afb_questionnairelie@odata.bind"?: string;
  afb_sourcedudepot: Afb_documentsafb_sourcedudepot;
  afb_statutdevalidite: Afb_documentsafb_statutdevalidite;
  afb_tailledufichierko?: number;
  "afb_tiers@odata.bind": string;
  afb_typededocument: string;
  afb_urlsharepoint: string;
  "afb_versionprecedente@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_documentsstatecode;
  statuscode?: Afb_documentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_documents extends Afb_documentsBase {
  afb_authentifiename?: string;
  afb_categoriename?: string;
  afb_questionnaireliename?: string;
  afb_sourcedudepotname?: string;
  afb_statutdevaliditename?: string;
  afb_tiersname?: string;
  afb_versionprecedentename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_categorie?: object;
  _afb_categorie_value?: string;
  afb_questionnairelie?: object;
  _afb_questionnairelie_value?: string;
  afb_tiers?: object;
  _afb_tiers_value?: string;
  afb_versionprecedente?: object;
  _afb_versionprecedente_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}